new commit

+++ src/main/java/com/paymentservice/dto/PayerDto.java
@@ -2,7 +2,7 @@ package com.paymentservice.dto;
 
 import java.time.LocalDate;
 
-public record PayerDto(Long payerId, String name, String secondName, LocalDate birthDate, String email, String phone) {
+public record PayerDto(Integer payerId, String name, String secondName, LocalDate birthDate, String email, String phone) {
 
 }
