+++ src/main/java/com/paymentservice/repository/PayerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PayerRepository extends JpaRepository<PayerEntity, Long> {
+public interface PayerRepository extends JpaRepository<PayerEntity, Integer> {
 }
